+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog 
 
+## 0.1.0
+
+- Added custom theme
+
+- Imported assets
+
+- Added custom widgets:
+  - Tile
+  - EntityCard
+  - EntityPage
+  - Dropdown
+  - ExpandedParagraph
+  - GeneratorPage
+
 ## 0.0.0
 
 - Started working on the app